requirements.txt
accelerate
diffusers==0.28.0
transformers==4.47.0
huggingface_hub==0.24.0
einops
decord
tqdm
safetensors
omegaconf
opencv-python
pydantic
compel
easydict
rotary_embedding_torch
imageio[ffmpeg]
gradio
httpx[socks]
numpy==1.26.3
hydra-core==1.1.1
hydra-colorlog
matplotlib
opencv-python
omegaconf
plotly
ftfy
torch==2.0.1
torchsde
torchdiffeq
vit-pytorch==1.4.4
pytorch-lightning==1.9.5
voltron-robotics
gitpython
scipy
sentence-transformers
torchsde
einops
torchdiffeq
gym
moviepy==1.0.3
tqdm
termcolor
wandb
einops_exts
sentencepiece 
mediapy
pyhash
lightning==2.4.0
piq==0.8.0
tensorflow==2.15.0
tensorflow_datasets==4.9.3
tensorflow_graphics==2021.12.3
jsonargparse[signatures]>=4.27.7
git+https://github.com/moojink/dlimp_openvla
